+++ dagger/src/main/java/boringyuri/dagger/DaggerModuleGeneratorStep.kt
@@ -22,8 +22,9 @@ import boringyuri.dagger.util.ProcessorOptions
 import boringyuri.processor.UriFactoryGeneratorStep.Companion.CONTAINER_IMPL_SUFFIX
 import boringyuri.processor.base.BoringProcessingStep
 import boringyuri.processor.base.ProcessingSession
+import boringyuri.processor.type.CommonTypeName
 import com.google.common.collect.ImmutableSet
-import com.google.common.collect.SetMultimap
+import com.google.common.collect.ImmutableSetMultimap
 import com.squareup.javapoet.ClassName
 import com.squareup.javapoet.MethodSpec
 import com.squareup.javapoet.TypeSpec
@@ -34,14 +35,14 @@ import javax.lang.model.util.ElementFilter
 
 class DaggerModuleGeneratorStep(session: ProcessingSession) : BoringProcessingStep(session) {
 
-    override fun annotations(): Set<Class<out Annotation>> {
-        return ImmutableSet.of(UriFactory::class.java)
+    override fun annotations(): Set<String> {
+        return ImmutableSet.of(UriFactory::class.java.name)
     }
 
     override fun process(
-        elementsByAnnotation: SetMultimap<Class<out Annotation>, Element>
+        elementsByAnnotation: ImmutableSetMultimap<String, Element>
     ): Set<Element> {
-        val factories = ElementFilter.typesIn(elementsByAnnotation[UriFactory::class.java])
+        val factories = ElementFilter.typesIn(elementsByAnnotation[UriFactory::class.java.name])
         val moduleName = ProcessorOptions.getModuleName(session)
 
         return generateBoringDaggerModule(moduleName, factories)
@@ -76,6 +77,7 @@ class DaggerModuleGeneratorStep(session: ProcessingSession) : BoringProcessingSt
         return MethodSpec.methodBuilder("provide${factoryName.simpleName()}")
             .addModifiers(Modifier.STATIC)
             .addAnnotation(DaggerTypeName.PROVIDES)
+            .addAnnotation(CommonTypeName.NON_NULL)
             .returns(factoryName)
             .addStatement("return new \$T()", factoryImplName)
             .build()