+++ processor/src/main/java/boringyuri/processor/TypeAdapterFactoryGeneratorStep.kt
@@ -24,7 +24,7 @@ import boringyuri.processor.type.CommonTypeName.*
 import boringyuri.processor.util.ProcessorOptions.getTypeAdapterFactory
 import com.google.auto.common.MoreTypes
 import com.google.common.collect.ImmutableSet
-import com.google.common.collect.SetMultimap
+import com.google.common.collect.ImmutableSetMultimap
 import com.squareup.javapoet.*
 import javax.lang.model.element.Element
 import javax.lang.model.element.Modifier
@@ -34,27 +34,42 @@ class TypeAdapterFactoryGeneratorStep(
     session: ProcessingSession
 ) : BoringProcessingStep(session) {
 
-    override fun annotations(): Set<Class<out Annotation>> {
-        return ImmutableSet.of(TypeAdapter::class.java)
+    private val typeAdapters = LinkedHashMap<ClassName, Element>()
+
+    override fun annotations(): Set<String> {
+        return ImmutableSet.of(TypeAdapter::class.java.name)
     }
 
     override fun process(
-        elementsByAnnotation: SetMultimap<Class<out Annotation>, Element>
+        elementsByAnnotation: ImmutableSetMultimap<String, Element>
     ): Set<Element> {
         // Stop processing Type Adapters if the factory class is not specified.
         // Every adapter instance will be created at use without caching.
         val typeAdapterFactory = getTypeAdapterFactory(session) ?: return emptySet()
 
-        val adaptableElements = elementsByAnnotation[TypeAdapter::class.java]
+        val adaptableElements = elementsByAnnotation[TypeAdapter::class.java.name]
 
-        val typeAdapters = adaptableElements.mapNotNull { element ->
+        if (typeAdapters.isNotEmpty()) {
+            adaptableElements.forEach {
+                logger.error(
+                    it,
+                    "%s is already created. Can't add %s to it.",
+                    typeAdapterFactory.simpleName(),
+                    it.simpleName
+                )
+            }
+
+            return emptySet() // early exit
+        }
+
+        adaptableElements.mapNotNull { element ->
             val typeAdapter = element.getAnnotation(TypeAdapter::class.java)?.valueMirror()
 
             // Map TypeAdapter class name to originating element to be able to
             // log an error or postpone element processing to the next step
             // in case it's possible.
             typeAdapter?.let { ClassName.get(MoreTypes.asTypeElement(it)) to element }
-        }.associate { it }
+        }.associateTo(typeAdapters) { it }
 
         return generateTypeAdapterFactory(typeAdapterFactory, typeAdapters)
     }