+++ sample/src/main/java/boringyuri/sample/uri/ContactUriBuilder.java
@@ -23,12 +23,15 @@ import android.net.Uri;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import boringyuri.api.matcher.MatchesTo;
 import boringyuri.api.Param;
 import boringyuri.api.Path;
 import boringyuri.api.UriBuilder;
 import boringyuri.api.UriFactory;
 import boringyuri.api.WithUriData;
+import boringyuri.api.matcher.WithUriMatcher;
 import boringyuri.api.adapter.TypeAdapter;
+import boringyuri.sample.BuildConfig;
 import boringyuri.sample.data.Address;
 import boringyuri.sample.data.adapter.RectTypeAdapter;
 
@@ -36,14 +39,24 @@ import boringyuri.sample.data.adapter.RectTypeAdapter;
         scheme = ContentResolver.SCHEME_CONTENT,
         authority = "boringyuri.sample.provider"
 )
+@WithUriMatcher("ContactUriMatcher")
 public interface ContactUriBuilder {
+    class Contract {
+        private static final String CONTACT_DATA = "CONTACT_DATA";
+        private static final String CONTACT_PHOTO = "CONTACT_PHOTO";
+        private static final String VCARD = "vcard";
+        private static final String HUAWEI_VCARD = "huawei_vcard";
+    }
+
     @NonNull
     @UriBuilder("/data/{contactId}")
+    @MatchesTo(value = Contract.CONTACT_DATA)
     Uri buildContactDataUri(@Path long contactId);
 
     @NonNull
     @UriBuilder("/file/photo/{group}/{contactId}")
     @WithUriData
+    @MatchesTo(value = Contract.CONTACT_PHOTO)
     Uri buildContactPhotoUri(
             @Path @NonNull String group,
             @Path long contactId,
@@ -51,11 +64,18 @@ public interface ContactUriBuilder {
 
     @NonNull
     @UriBuilder("/file/vcard/{contactId}")
+    @WithUriData
+    @MatchesTo(Contract.VCARD)
     Uri buildVCardUri(
             @Path long contactId,
             @NonNull @Param String firstName,
             @Nullable @Param String lastName,
             @Nullable @Param("address") Address homeAddress);
+
+    @NonNull
+    @UriBuilder("/file/huawei/{contactId}")
+    @MatchesTo(value = Contract.HUAWEI_VCARD, enabled = BuildConfig.NO_PLAY_SERVICES)
+    Uri buildNoPlayServicesVCard(@Path long contactId);
 
     @NonNull
     static ContactUriBuilder create() {